requirements.txt
tqdm==4.55.0
GitPython==3.1.24
travispy==0.3.5
PyGithub
loguru